requirements.txt
Flask==2.0.1
flask-cors==3.0.10
flask-sqlalchemy==3.0.2
python-jose==3.3.0
passlib==1.7.4
textblob==0.15.3
gunicorn==20.1.0
psycopg2-binary==2.9.3
python-dotenv==0.19.2
requests==2.26.0
nltk==3.6.3
pytest==6.2.5
pytest-cov==2.12.1
prometheus-flask-exporter==0.23.2 